requirements.txt
MouseInfo==0.1.3
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyirsdk==1.2.6
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
PyYAML==5.3.1
